1. Вместо Date использован LocalDate 2. Избавился от метода с подчеркиванием 3. В методе получения строк с вопросами вынесена в отдельный метод логика создания фильтра 4. Изменен метод обновения вопроса (избавился от удалений) 5. Добавил возможжность дробного процента 6. Исправил сокрытие верных ответов от пользователя 7. Немного поменял тест на расчет процента, добавил коментариев и assert-ов

+++ unit-11/app-server/src/main/java/ru/kitzsable/appserver/transfer/JournalLineResponseDTO.java
@@ -6,10 +6,10 @@ import java.util.List;
  * DTO Ответа на запрос строк журнала
  */
 public class JournalLineResponseDTO {
-    public int total;
+    public Long total;
     public List<? extends JournalLineDTO> items;
 
-    public JournalLineResponseDTO(int total, List<? extends JournalLineDTO> item) {
+    public JournalLineResponseDTO(Long total, List<? extends JournalLineDTO> item) {
         this.total = total;
         this.items = item;
     }
